Adding test class SynchronousSearchSvcImplTest.
